Change to correct checksum
The only documentation I could find online on how the checksum is
calculated said it's the one's compliment of the sum of all bytes
transmitted in the message so far. After capturing the transactions from
existing ALDL software, it seems that it should actually by the two's
compliment.

This still needs to be tested though... again, a job for tomorrow, when
it is light out.